Fix local map resize type and format copyMapRegion

+++ perception/autoware_probabilistic_occupancy_grid_map/lib/costmap_2d/occupancy_grid_map_base.cpp
@@ -149,7 +149,7 @@ void OccupancyGridMapInterface::updateOrigin(double new_origin_x, double new_ori
     return;
 #endif
   } else {
-    local_map.resize(static_cast<size_t>(cell_size_x) * cell_size_y);
+    local_map.resize(static_cast<size_t>(cell_size_x) * static_cast<size_t>(cell_size_y));
 
     // copy the local window in the costmap to the local map
     copyMapRegion(
